+++ src/input/examples/append-prepend/append-prepend.component.ts
@@ -1,0 +1,11 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'thy-input-append-prepend-example',
+    templateUrl: './append-prepend.component.html'
+})
+export class ThyInputAppendPrependExampleComponent implements OnInit {
+    constructor() {}
+
+    ngOnInit() {}
+}