+++ src/layout/layout.module.ts
@@ -11,6 +11,8 @@ import { ThyIconModule } from 'ngx-tethys/icon';
 import { ThyTooltipModule } from 'ngx-tethys/tooltip';
 import { ThySidebarHeaderComponent } from './sidebar-header.component';
 import { ThySidebarFooterComponent } from './sidebar-footer.component';
+import { ThySidebarContentComponent } from './sidebar-content.component';
+import { ThyResizableModule } from 'ngx-tethys/resizable';
 
 @NgModule({
     declarations: [
@@ -19,17 +21,19 @@ import { ThySidebarFooterComponent } from './sidebar-footer.component';
         ThyContentComponent,
         ThySidebarComponent,
         ThySidebarHeaderComponent,
+        ThySidebarContentComponent,
         ThySidebarFooterComponent,
         ThyContentSectionComponent,
         ThyContentMainComponent
     ],
-    imports: [CommonModule, DragDropModule, ThyIconModule, ThyTooltipModule],
+    imports: [CommonModule, DragDropModule, ThyIconModule, ThyTooltipModule, ThyResizableModule],
     exports: [
         ThyLayoutComponent,
         ThyHeaderComponent,
         ThyContentComponent,
         ThySidebarComponent,
         ThySidebarHeaderComponent,
+        ThySidebarContentComponent,
         ThySidebarFooterComponent,
         ThyContentSectionComponent,
         ThyContentMainComponent