+++ src/menu/group/menu-group.component.ts
@@ -1,8 +1,20 @@
-import { Component, OnInit, HostBinding, Input, Output, EventEmitter, ElementRef, ViewChild, TemplateRef } from '@angular/core';
+import {
+    Component,
+    OnInit,
+    HostBinding,
+    Input,
+    Output,
+    EventEmitter,
+    ElementRef,
+    ViewChild,
+    TemplateRef,
+    ContentChild
+} from '@angular/core';
 import { trigger, state, style, transition, animate } from '@angular/animations';
 import { ComponentType } from '@angular/cdk/portal';
 import { ThyPopover } from 'ngx-tethys/popover';
 import { ThyMenuComponent, ThyMenuTheme } from '../menu.component';
+import { InputBoolean } from 'ngx-tethys/core';
 
 @Component({
     selector: 'thy-menu-group, [thy-menu-group],[thyMenuGroup]',
@@ -47,11 +59,14 @@ export class ThyMenuGroupComponent implements OnInit {
 
     @ViewChild('thyMenuGroup', { static: true }) _thyMenuGroup: ElementRef;
 
+    @ContentChild('headerContent')
+    headerContentTemplateRef: TemplateRef<any>;
+
     @HostBinding('class.thy-menu-group') isThyMenuGroup = true;
 
     @HostBinding('class.has-icon') showIcon = false;
 
-    @HostBinding('class.collapsed') isCollapsed = true;
+    @HostBinding('class.collapsed') isCollapsed = false;
 
     @Input() thyTitle = '';
 
@@ -59,6 +74,14 @@ export class ThyMenuGroupComponent implements OnInit {
     set thyExpand(value: boolean) {
         this.isCollapsed = !!!value;
     }
+
+    @Input('thyCollapsed')
+    @InputBoolean()
+    set thyCollapsed(value: boolean) {
+        this.isCollapsed = value;
+    }
+
+    @Input() @InputBoolean() thyCollapsible: boolean = true;
 
     @Input('thyShowIcon')
     set thyShowIcon(value: boolean) {
@@ -70,17 +93,12 @@ export class ThyMenuGroupComponent implements OnInit {
         this.iconClass = value;
     }
 
-    @Input() thyCanArrow: boolean = true;
-
     @Input('thyActionIcon')
     set thyActionIcon(value: string) {
         this.rightIconClass = value;
     }
 
     @Input() thyShowAction = false;
-
-    @Input('thyHeaderContent')
-    headerContentTemplate: TemplateRef<any>;
 
     @Input() thyActionStopPropagation = true;
 
@@ -96,7 +114,7 @@ export class ThyMenuGroupComponent implements OnInit {
     ngOnInit(): void {}
 
     collapseGroup(): void {
-        if (!this.thyCanArrow) {
+        if (!this.thyCollapsible) {
             return;
         }
         this.isCollapsed = !this.isCollapsed;