+++ src/space/space.component.ts
@@ -32,14 +32,16 @@ export class ThySpaceItemDirective implements OnInit {
     ngOnInit(): void {}
 }
 
-type Size = 'zero' | 'xs' | 'sm' | 'md' | 'lg' | number;
+type Size = 'zero' | 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xlg' | number;
 
 const SIZE_SPACE_MAP = {
     zero: 0,
-    xs: 5,
-    sm: 10,
-    md: 15,
-    lg: 20
+    xxs: 4,
+    xs: 8,
+    sm: 12,
+    md: 16,
+    lg: 20,
+    xlg: 24
 };
 
 const DEFAULT_SIZE: Size = 'md';
@@ -62,7 +64,7 @@ export class ThySpaceComponent extends _MixinBase implements OnInit, AfterConten
     public space: number = SIZE_SPACE_MAP[DEFAULT_SIZE];
 
     /**
-     * 大小，支持 'zero' | 'xs' | 'sm' | 'md' | 'lg' 和自定义数字大小
+     * 大小，支持 'zero' | 'xxs' | 'xs' | 'sm' | 'md' | 'lg' | 'xlg' 和自定义数字大小
      * @type string | number
      */
     @Input() set thySize(size: Size) {